react refactor pass 1
